pyproject.toml
[project]
name = "skills-scanner-skill"
version = "0.1.0"
description = "A GitHub Copilot skill that scans AI agent skills for prompt injection, data exfiltration, and malicious code patterns"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "openai-agents>=0.3.2",
    "openai>=1.109.1",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.9.0",
    "yara-python>=4.5.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "typer>=0.15.0",
    "pyyaml>=6.0.2",
    "aiofiles>=24.1.0",
    "rich>=13.9.0",
]

[project.scripts]
skills-scanner = "main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/scanner"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
